+++ src/elements/navigation/package.json
@@ -14,8 +14,10 @@
     "react": "^16.7.0"
   },
   "dependencies": {
+    "@uswitch/trustyle.button": "^3.0.4",
+    "@uswitch/trustyle.button-link": "^3.0.4",
     "@uswitch/trustyle.icon": "^1.19.1",
-    "@uswitch/trustyle.tabs": "^1.0.9",
-    "@uswitch/trustyle.imgix-image": "^0.1.40"
+    "@uswitch/trustyle.imgix-image": "^0.1.40",
+    "@uswitch/trustyle.tabs": "^1.0.9"
   }
 }
